Reduce parser duplication
